// src/modules/autorizacion/enums/autorizacion.enums.ts
/**
 * Enums para el módulo de autorización
 */

export enum AccionHttp {
  GET = 'GET',
  POST = 'POST',
  PUT = 'PUT',
  DELETE = 'DELETE',
  PATCH = 'PATCH',
}

export enum AplicacionTipo {
  BACKEND = 'backend',
  FRONTEND = 'frontend',
}

export enum EstadoRol {
  ACTIVO = 'activo',
  INACTIVO = 'inactivo',
}

export enum EstadoUsuarioRol {
  ACTIVO = 'activo',
  INACTIVO = 'inactivo',
}